pyproject.toml
[project]
name = "autoreduce"
version = "0.1.0"
description = "Autoreduce - simplify code while preserving correctness"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "matplotlib>=3.8",
    "pandas>=2.1",
    "pytest>=7.0",
    "ruff>=0.15.6",
]

[project.optional-dependencies]
# Optional: richer complexity analysis
analysis = [
    "radon>=6.0",           # Python cyclomatic complexity (more accurate than AST walk)
    "lizard>=1.17",         # Multi-language complexity analysis
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.ruff.lint.per-file-ignores]
"run.py" = ["E402"]
"tests/*.py" = ["E402"]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
